backend/requirements.txt
bcrypt
fastapi
numpy
python-dotenv
google-genai
google-generativeai
chromadb
pypdf
langgraph
supabase
uvicorn
pydantic[email]